Fix TrkQual loop.  Convert to ParticleTrajectory in KalSeed interface
