+++ EventDisplay/src/DataInterface.cc
@@ -790,9 +790,9 @@ void DataInterface::findBoundaryP(spaceminmax &m, double x, double y, double z)
   if(std::isnan(m.maxz) || z>m.maxz) m.maxz=z;
 }
 
-using LHPT = KinKal::PiecewiseTrajectory<KinKal::LoopHelix>;
-using CHPT = KinKal::PiecewiseTrajectory<KinKal::CentralHelix>;
-using KLPT = KinKal::PiecewiseTrajectory<KinKal::KinematicLine>;
+using LHPT = KinKal::ParticleTrajectory<KinKal::LoopHelix>;
+using CHPT = KinKal::ParticleTrajectory<KinKal::CentralHelix>;
+using KLPT = KinKal::ParticleTrajectory<KinKal::KinematicLine>;
 template<class KTRAJ> void DataInterface::fillKalSeedTrajectory(std::unique_ptr<KTRAJ> &trajectory,
                                                                 int particleid, int trackclass, int trackclassindex, double p1,
                                                                 boost::shared_ptr<ComponentInfo> info)
